+++ packages/ui-components/src/chat/chat.stories.tsx
@@ -84,7 +84,7 @@ export const Chat = (): JSX.Element => {
     const [value, setValue] = React.useState(messages);
 
     return (
-        <div style={{ backgroundColor: 'black', display: 'flex' }}>
+        <div style={{ backgroundColor: 'white', display: 'flex', height: '100%' }}>
             <ChatComponent activeUser={Harry} onUpdate={setValue} value={value} />
         </div>
     );